IG60: Add support for USB local update in igupd
USB local update uses device service to give the status of the update.
If the USB local update fails, it will blink red device led until usb
is not removed. And, the local update service will switch back to
SURICATTA mode. The swupdate.swu package is a signed package which is
verified by the swupdate tool with the certificate on the device.

bug 13445