Include secrets in the developer's build.
In order for swupdate to work properly, we need to have the keys present
on the system when swupdate is running.  We can't generate the read only
encrypted filesystem as part of the build because elevated permissions are
necessary to mount and create the encrypted fs.  The keys and .ext4 fill
will remain static and modified and pushed the repo when necessary. In
development, the same keys are used on the swu images as are used to
sign and store in the dtbs.

Bug 13784