Secure boot verification test does not validate rodata
The IG Update Service (igupd) displays the "Secure Boot Cycle Complete"
message, which is used by the production process (WBX3) to validate that
the secure programming process was properly executed.  However,
inspection of the code shows that this message only relies on validating
the public signing certificate in the root filesystem
(/etc/ssl/misc/dev.crt), and does not validate any data on the encrypted
factory data (/rodata).  Thus, this message could appear in a
"false positive" where rodata is not in place or cannot be decrypted.

Bug 16255